+++ src/main/java/main/Parser.java
@@ -1,10 +1,10 @@
 package main;
 
 
-import task.deadline.Deadline;
-import task.event.Event;
+import task.Deadline;
+import task.Event;
 import task.Task;
-import task.Todo.Todo;
+import task.Todo;
 import java.util.ArrayList;
 
 /**
