+++ src/main/java/main/Storage.java
@@ -1,8 +1,8 @@
 package main;
-import task.deadline.Deadline;
-import task.event.Event;
+import task.Deadline;
+import task.Event;
 import task.Task;
-import task.Todo.Todo;
+import task.Todo;
 
 import java.io.File;
 import java.io.FileNotFoundException;
