+++ src/main/java/task/Event.java
@@ -1,28 +1,24 @@
-package task.deadline;
+package task;
 
 import task.Task;
 
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 
 /**
- * Represents a deadline under Task class with date-time format
+ * Represents an event under Task class
  * Allows users to get_type and get_task
  */
-public class Deadline extends Task {
+public class Event extends Task {
     private String type;
-    private String task = "";
-    private LocalDate dateTime;
-    private String dateTime2;
-
+    private String date_time = "";
+    private String task;
 
     /**
-     * Constructor to create a deadline task for newly inputed messages
+     * Constructor to create an event task for newly inputed messages
      * Sieve out the task and date separately
      *
      * @param message String the user input message
      */
-    public Deadline (String message) {
+    public Event (String message) {
         super(message);
         this.setDateTime();
         this.setTask();
@@ -30,14 +26,15 @@ public class Deadline extends Task {
     }
 
 
+
     /**
-     * Constructor to create a deadline task for Duke.txt lines
+     * Constructor to create an event task for Duke.txt lines
      * Sieve out the task and date separately
      *
      * @param message String the lines in Duke.txt file
      * @param isDuke Boolean to differentiate between 2 constructors, always true
      */
-    public Deadline (String message, boolean isDuke) {
+    public Event (String message, boolean isDuke) {
         super(message);
         this.setDateTime2();
         this.setTask2();
@@ -47,11 +44,6 @@ public class Deadline extends Task {
     @Override
     public String getType() {
         return this.type;
-    }
-
-    @Override
-    public void setType() {
-        this.type = "D";
     }
 
     /**
@@ -64,8 +56,8 @@ public class Deadline extends Task {
         int startIndex = 0;
         int endIndex = 0;
         for (int i = 0; i < this.message.length(); i++) {
-            if (this.message.substring(i, i+1).equals("d")) {
-                startIndex = i + 8;
+            if (this.message.substring(i, i+1).equals("t")) {
+                startIndex = i + 2;
                 break;
             }
         }
@@ -75,15 +67,14 @@ public class Deadline extends Task {
                 break;
             }
         }
-        this.task = message.substring(startIndex,endIndex)
-                + " (by "
-                    + this.getDateTime()
-                        + ")";
+        this.task = " "
+                + message.substring(startIndex,endIndex)
+                    + this.getDateTime();
     }
 
     /**
      * Sieve out the task + date portion of the message
-     * Only for lines in Duke.txt
+     * Only for Duke.txt lines
      *
      */
     @Override
@@ -91,8 +82,8 @@ public class Deadline extends Task {
         int startIndex = 0;
         int endIndex = 0;
         for (int i = 0; i < this.message.length(); i++) {
-            if (this.message.substring(i, i+1).equals("d")) {
-                startIndex = i + 9;
+            if (this.message.substring(i, i+1).equals("e")) {
+                startIndex = i + 6;
                 break;
             }
         }
@@ -103,17 +94,22 @@ public class Deadline extends Task {
             }
         }
         this.task = message.substring(startIndex,endIndex)
-                + " "
-                    + this.getDateTime2();
+                + this.getDateTime();
     }
+
 
     @Override
     public String getTask() {
         return this.task;
     }
 
+    @Override
+    public void setType() {
+        this.type = "E";
+    }
+
     /**
-     * Sieve out the date portion of the message as LocalDate object
+     * Sieve out the date portion of the message
      * Only for newly inputed message
      *
      */
@@ -126,36 +122,30 @@ public class Deadline extends Task {
                 break;
             }
         }
-        String dateTiming = message.substring(startIndex);
-        this.dateTime = LocalDate.parse(dateTiming);
+        this.date_time = " (at "
+                + message.substring(startIndex)
+                    + ")";
     }
 
 
     /**
      * Sieve out the date portion of the message
-     * Only for Duke.txt messages
+     * Only for Duke.txt lines
      *
      */
     public void setDateTime2() {
-        int length = this.message.length();
-        this.dateTime2 = "("
-                + message.substring(length - 15);
+        int startIndex = 0;
+        for (int i = 0; i < this.message.length(); i++) {
+            if (this.message.substring(i, i+1).equals("(")) {
+                startIndex = i - 1;
+                break;
+            }
+        }
+        this.date_time = message.substring(startIndex);
     }
 
-
-    /**
-     * Print out the LocalDate object as month date year format
-     * Only for newly inputed messages
-     *
-     * @return date string
-     *
-     */
+    @Override
     public String getDateTime() {
-        return this.dateTime.format(DateTimeFormatter.ofPattern("MMM d yyyy"));
+        return date_time;
     }
-
-    public String getDateTime2() {
-        return this.dateTime2;
-    }
-
 }
