+++ src/main/java/task/Todo.java
@@ -1,4 +1,4 @@
-package task.Todo;
+package task;
 
 import task.Task;
