+++ src/test/java/EventTest.java
@@ -1,6 +1,6 @@
-package task.event;
-
 import org.junit.jupiter.api.Test;
+import task.Event;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
     public class EventTest {
