+++ src/test/java/TodoTest.java
@@ -1,6 +1,6 @@
-package task.Todo;
-
 import org.junit.jupiter.api.Test;
+import task.Todo;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TodoTest {
